Add binutils-mesboot0:
I had many problems making the buildsystem make it and install it
properly. Mostly, Make was trying to run in bfd/po and as there's no
Makefile there it exploded. We patched around that making empty files
with the name of the problematic targets.

`objdump`, `strip` and some more work, but `ar` and `as` are broken.

- `as`: Fails all the time, the only thing it can do is `--help`
- `ar`: Fails when archiving many files but works with few. Probably not
  enough for GCC...